# api/openapi.yaml
openapi: 3.0.3
info:
  title: VenueOS API
  version: 1.0.0
  description: Initial contract for the VenueOS backend. The user endpoints are the transitional template slice and will evolve through the documented implementation roadmap.
servers:
  - url: /
paths:
  /healthz:
    get:
      operationId: Healthz
      summary: Check process liveness
      responses:
        "204":
          description: The process is alive.
  /readyz:
    get:
      operationId: Readyz
      summary: Check dependency readiness
      responses:
        "204":
          description: The service can receive traffic.
        "503":
          $ref: "#/components/responses/Problem"
  /v1/users:
    get:
      operationId: ListUsers
      summary: List users
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "400":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
    post:
      operationId: CreateUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: User created.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /v1/users/{id}:
    get:
      operationId: GetUser
      summary: Fetch a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The requested user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
components:
  responses:
    Problem:
      description: An RFC 9457 problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CreateUser:
      type: object
      additionalProperties: false
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        name:
          type: string
          minLength: 1
          maxLength: 100
    User:
      type: object
      additionalProperties: false
      required: [id, email, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserList:
      type: object
      additionalProperties: false
      required: [items, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        limit:
          type: integer
        offset:
          type: integer
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: object
          additionalProperties:
            type: string